Restructured linalg kernels into a folder, changed imports accordingly
