+++ src/qttools/kernels/linalg/__init__.py
@@ -1,0 +1,14 @@
+# Copyright (c) 2024 ETH Zurich and the authors of the qttools package.
+
+
+from qttools.kernels.linalg.eig import eig
+from qttools.kernels.linalg.eigvalsh import eigvalsh
+from qttools.kernels.linalg.qr import qr
+from qttools.kernels.linalg.svd import svd
+
+__all__ = [
+    "eig",
+    "svd",
+    "eigvalsh",
+    "qr",
+]
