+++ src/qttools/kernels/linalg/eig.py
