+++ src/qttools/kernels/linalg/eigvalsh.py
