+++ src/qttools/kernels/linalg/qr.py
