+++ src/qttools/kernels/linalg/svd.py
